backend/requirements/testing.txt
-r base.txt

pytest==6.2.2
pytest-django==4.1.0
pytest-mock==3.5.1